+++ src/ai_lib_python/utils/__init__.py
@@ -8,6 +8,12 @@ This module contains:
 - Other helper functions
 """
 
+from ai_lib_python.utils.thinking_extract import (
+    OPENAI_COMPAT_THINKING_KEYS,
+    first_nonempty_string_field,
+    thinking_from_openai_compat_delta,
+    thinking_from_openai_compat_message,
+)
 from ai_lib_python.utils.tool_call_assembler import (
     MultiToolCallAssembler,
     ToolCallAssembler,
@@ -16,6 +22,10 @@ from ai_lib_python.utils.tool_call_assembler import (
 
 __all__ = [
     "MultiToolCallAssembler",
+    "OPENAI_COMPAT_THINKING_KEYS",
     "ToolCallAssembler",
     "ToolCallFragment",
+    "first_nonempty_string_field",
+    "thinking_from_openai_compat_delta",
+    "thinking_from_openai_compat_message",
 ]